chore: Update docker-compose.yml to fix user permissions
This commit updates the docker-compose.yml file to fix the user permissions for the chat-mongodb service. Previously, the user was set as "$1000:1000", which caused an error. The user is now set as "1000:1000" to ensure proper permissions. This change resolves the issue with user permissions and allows the service to run without errors.